add mdi remark plugin
